remove all references to Dataset
