+++ robustnessgym/ops/similarity.py
@@ -5,7 +5,7 @@ import cytoolz as tz
 import numpy as np
 from datasets import load_metric
 
-from robustnessgym.core.dataset import transpose_batch
+from robustnessgym.core.tools import transpose_batch
 from robustnessgym.core.operation import Operation, lookup
 from robustnessgym.core.slice import SliceDataPanel as DataPanel
 from robustnessgym.ops.spacy import SpacyOp
