+++ tests/ops/test_spacy.py
