+++ tests/ops/test_stanza.py
